# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation for Queo Challenge",
        "contact": {
            "name": "Admin",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/validate": {
            "post": {
                "tags": [
                    "Validation"
                ],
                "summary": "Validar respuesta al desafío",
                "description": "Valida la respuesta proporcionada para el desafío actual. Limitado a 1 intento por minuto.",
                "operationId": "validateChallenge",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "number_of_groups",
                                    "answer"
                                ],
                                "properties": {
                                    "number_of_groups": {
                                        "description": "Número total de grupos encontrados",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "answer": {
                                        "description": "Nombres de los grupos separados por coma",
                                        "type": "string",
                                        "example": "Group A, Group B"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Respuesta validada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "correct"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "¡Tu respuesta es correcta!"
                                                },
                                                "timestamp": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-10T06:31:42Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token inválido o expirado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "La respuesta es incorrecta. Por favor, revisa los nombres de los grupos."
                                        },
                                        "errors": {
                                            "properties": {
                                                "number_of_groups": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El número de grupos no coincide con la solución."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiadas solicitudes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Has excedido el límite de intentos. Por favor, espera 60 segundos."
                                        },
                                        "limits": {
                                            "properties": {
                                                "max_attempts": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "attempts_left": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "retry_after_seconds": {
                                                    "type": "integer",
                                                    "example": 60
                                                },
                                                "retry_after_minutes": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error inesperado durante la validación."
                                        },
                                        "reference_id": {
                                            "type": "string",
                                            "example": "err_5f3e9b2c1d"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}